+++ Changelog.md
@@ -7,6 +7,7 @@ Breaking Changes for 3.0.0:
 Enhancements
 
 * Clean up some internal use of Enumerable methods. (Vipul A M)
+* Replace unmaintained syntax gem with coderay gem. (Xavier Shay)
 
 Bug fix